[Milestone 1] Restructured project, Added method to generate response body
